requirements.txt
fastapi
uvicorn
pydantic
python-jose[cryptography]
requests
bcrypt
python-multipart
